# apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "app/dynamic_data/api/data.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/gen-code-service/api/gen-react-code": {
      "post": {
        "operationId": "GenCodeService_GenReactSourceCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGenReactSourceCodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGenReactSourceCodeRequest"
            }
          }
        ],
        "tags": [
          "GenCodeService"
        ]
      }
    },
    "/gen-code-service/api/hello-world": {
      "get": {
        "operationId": "GenCodeService_HelloWorld",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiHelloWorldResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "GenCodeService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiGenReactSourceCodeRequest": {
      "type": "object",
      "properties": {
        "nodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNode"
          }
        },
        "pages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiPage"
          }
        }
      }
    },
    "apiGenReactSourceCodeResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    },
...
    "apiHelloWorldResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "apiNode": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "props": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "hidden": {
          "type": "boolean"
        },
        "children": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pagePath": {
          "type": "string"
        }
      }
    },
    "apiPage": {
      "type": "object",
      "properties": {
        "path": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}